# docs/swagger/facebook-ads-swagger.yaml
openapi: 3.0.3
info:
  title: Marketing Hub - Facebook Ads API
  version: 1.0.0
  description: |
    Contrato dos endpoints do backend que atendem a integração com Facebook Ads.
    A implementação canônica dos controllers fica concentrada no pacote
    `com.marketinghub.facebookads.controller`, mantendo as rotas públicas legadas
    para compatibilidade com frontend e workers.
...
paths:
...
  /api/experiments/{experimentId}/facebook-campaigns:
    get:
      tags: [Facebook Campaigns]
      summary: Lista campanhas Facebook vinculadas ao experimento.
      parameters: [ { name: experimentId, in: path, required: true, schema: { type: integer, format: int64 } } ]
      responses: { '200': { description: Campanhas retornadas. } }
  /api/experiments/{experimentId}/facebook-campaigns/reset-preview:
    get:
      tags: [Facebook Campaigns]
      summary: Simula reset das campanhas Facebook do experimento.
      parameters: [ { name: experimentId, in: path, required: true, schema: { type: integer, format: int64 } } ]
      responses: { '200': { description: Prévia retornada. } }
  /api/experiments/{experimentId}/facebook-campaigns/reset:
    post:
      tags: [Facebook Campaigns]
      summary: Executa reset das campanhas Facebook do experimento.
      parameters: [ { name: experimentId, in: path, required: true, schema: { type: integer, format: int64 } } ]
      responses: { '200': { description: Reset executado. } }
...
  /api/experiments/{experimentId}/adset-playbook:
    get:
      tags: [Facebook Playbook]
      summary: Consulta workflow do playbook de adsets por experimento.
      parameters: [ { name: experimentId, in: path, required: true, schema: { type: integer, format: int64 } } ]
      responses: { '200': { description: Workflow retornado. } }
  /api/experiments/{experimentId}/adset-playbook/start:
    post:
      tags: [Facebook Playbook]
      summary: Inicia workflow do playbook de adsets.
      parameters: [ { name: experimentId, in: path, required: true, schema: { type: integer, format: int64 } } ]
      responses: { '202': { description: Workflow iniciado. } }
  /api/experiments/{experimentId}/adset-playbook/jobs/{jobId}:
    get:
      tags: [Facebook Playbook]
      summary: Consulta detalhes de job do playbook de adsets.
      parameters:
        - { name: experimentId, in: path, required: true, schema: { type: integer, format: int64 } }
        - { name: jobId, in: path, required: true, schema: { type: integer, format: int64 } }
      responses: { '200': { description: Job retornado. } }
...